added components inspector

+++ Pearly/src/Platform/OpenGL/OpenGLShader.cpp
@@ -204,17 +204,17 @@ namespace Pearly {
 		std::unordered_map<GLenum, std::string> shaderSources;
 
 		const char* typeToken = "#type";
-		uint32 typeTokenLength = strlen(typeToken);
+		size_t typeTokenLength = strlen(typeToken);
 		size_t pos = source.find(typeToken, 0);
 		while (pos != std::string::npos)
 		{
-			uint32 endOfLine = source.find_first_of("\r\n", pos);
+			size_t endOfLine = source.find_first_of("\r\n", pos);
 			PR_CORE_ASSERT(endOfLine != std::string::npos, "Syntax error!");
-			uint32 begin = pos + typeTokenLength + 1;
+			size_t begin = pos + typeTokenLength + 1;
 			std::string type = source.substr(begin, endOfLine - begin);
 			PR_CORE_ASSERT(ShaderTypeFromString(type), "Invalid shader type specified '{0}'!", type);
 
-			uint32 nextLinePos = source.find_first_not_of("\r\n", endOfLine);
+			size_t nextLinePos = source.find_first_not_of("\r\n", endOfLine);
 			PR_CORE_ASSERT(nextLinePos != std::string::npos, "Syntax error");
 			pos = source.find(typeToken, nextLinePos);
 			shaderSources[ShaderTypeFromString(type)] = (pos == std::string::npos) ? source.substr(nextLinePos) : source.substr(nextLinePos, pos - nextLinePos);
